add step 11

+++ api-suivi-eleves-etml/app/controllers/comments_controller.ts
@@ -2,7 +2,7 @@ import type { HttpContext } from '@adonisjs/core/http'
 import { commentValidator } from '#validators/comment'
 import Comment from '#models/comment'
 import Student from '#models/student'
-
+import CommentPolicy from '#policies/comment_policy'
 export default class CommentsController {
   /**
    * Afficher tous les commentaires de l'élève student_id
@@ -20,19 +20,24 @@ export default class CommentsController {
   /**
    * Ajouter un nouveau commentaire à l'élève student_id
    */
-  async store({ params, request, response }: HttpContext) {
+  async store({ params, request, response, auth }: HttpContext) {
     // Récupération des données envoyées par le client et validation des données
     const { content } = await request.validateUsing(commentValidator)
-    // Pour l'instant, on fixe "en dur" l'id de l'enseignant
-    // Lorsque l'on va mettre en place l'authentification,
-    // on pourra récupérer l'id de l'enseignant connecté
-    const teacherId = 1
+    // Récupération de l'utilisateur authentifié
+    const user = auth.user!
+    // Chargement de l'enseignant lié à cet utilisateur
+    const teacher = await user.related('teacher').query().first()
+    if (!teacher) {
+      return response.badRequest({ message: 'Teacher not found' })
+    }
+    const teacherId = teacher.id
     // Création du commentaire lié à l'élève
     const comment = await Comment.create({
       content,
       studentId: params.student_id,
       teacherId,
     })
+    // Réponse HTTP 201 avec le commentaire
     return response.created(comment)
   }
   /**
@@ -51,7 +56,7 @@ export default class CommentsController {
   /**
    * Mettre à jour le commentaire de l'élève student_id
    */
-  async update({ params, request, response }: HttpContext) {
+  async update({ params, request, response, bouncer }: HttpContext) {
     // Récupération des données envoyées par le client et validation des données
     const { content } = await request.validateUsing(commentValidator)
     // Vérifie que le commentaire appartient bien à l'élève
@@ -59,6 +64,15 @@ export default class CommentsController {
       .where('id', params.id)
       .where('student_id', params.student_id)
       .firstOrFail()
+
+    // Vérifie les permissions de l'utilisateur connecté
+    if (await bouncer.with(CommentPolicy).denies('update', comment)) {
+      return response.unauthorized({
+        message:
+          "Vous n'êtes pas l'auteur de ce commentaire. Vous n'avez pas le droit de le modifier",
+      })
+    }
+
     // Mise à jour
     comment.content = content
     await comment.save()
@@ -68,11 +82,19 @@ export default class CommentsController {
   /**
    * Supprimer le commentaire de l'élève student_id
    */
-  async destroy({ params, response }: HttpContext) {
+  async destroy({ params, response, bouncer }: HttpContext) {
     const comment = await Comment.query()
       .where('id', params.id)
       .where('student_id', params.student_id)
       .firstOrFail()
+
+    // Vérifie les permissions de l'utilisateur connecté
+    if (await bouncer.with(CommentPolicy).denies('delete', comment)) {
+      return response.unauthorized({
+        message:
+          "Vous n'êtes pas l'auteur de ce commentaire. Vous n'avez pas le droit de le supprimer",
+      })
+    }
     // Suppression du commentaire
     await comment.delete()
     // On utilise `response.noContent` pour retourner un code HTTP 204 sans contenu
